refactor: Update chat component to display user avatars dynamically
The code changes in the chat component HTML and SCSS files update the avatar display to dynamically show user avatars based on the number of users. The previous static avatar images have been replaced with a loop that iterates over the users array and displays the profile images. This improves the flexibility and scalability of the chat component.