Update: FirebaseService & private-message component have been further developed
Functions have been implemented in the FirebaseService that return user-specific data to display it dynamically. These functions are needed in several places in the project. This makes it possible to keep the code small and easy to maintain. Furthermore, the HTML part of the private message component has been adapted to display the user chat data dynamically.