Update: FirebaseService & create-channel Component
A function has been added to the Firebase service that checks whether a specific channel name already exists in the collection. If the name already exists, no new document can be created. It is also not possible to give an existing channel an identical name when editing. This is required in several places within the project. Furthermore, a corresponding error message has been displayed in the create-channel component.