Update: implemented form validation & err msgs for current-user-profile
A form validation has been implemented which checks whether the edited name consists of at least 2 words (first and last name) as well as validating the edited email address. The corresponding error messages are displayed and it is not possible to submit the form. The design has been further adapted.