refafactor: data within the chat-histroy dynamically displayed & filter logic implemented
A logic has been implemented in the chat history template which differentiates between user & currentUser msg. The data is now displayed dynamically for name / avatar / time / date. For the date, a logic must be implemented afterwards which decides whether the msg is in the past, in which case the date should be visible, or whether the msg was written on the same day, in which case "today" should be displayed instead of the date.