Update: create new privateChat function / logic implemented - prvtChat model adjusted
The privateChat model has been adjusted. A function has been implemented in the user profile component which first creates a new private chat between the current user and the user from the displayed user profile. The functions for this are in the chat service. Functions have been implemented there which create a new privateChat doc, fill it with the existing data and, after creation, update the privateChatId field with the docref. When creating a new private chat document, logic is subsequently implemented so that it can be accessed from the new-message component. From there, a first message is sent directly when it is created.