Update / Refactor: message model adjusted bridge build between shared chat input & prvt chat componente
The message model has been adjusted. Functions have been implemented in the chat service. When a message is sent from the shared chat input, a distinction is first made as to which component it comes from. A new message is then created in the defined collection/subcolletion "messages". After the message doc has been created, it is updated with the doc ref. In addition, a helper function has been implemented that converts the time stored as a string in milliseconds into a defined format "HH:MM clock". This allows the time to be displayed in the desired format.