refactor: Update "Add Members to Channel" component
This commit refactors the "Add Members to Channel" component by updating the HTML template and SCSS file. The changes include:
- Updating the close button to use a div element instead of a button element
- Adding width and height styles to the close button image
- Adding ngModel binding to the input element for the user name
- Enabling the "Hinzufügen" button only when the user name is not empty

This commit also includes changes to the TypeScript file, where a new method `closeWindow()` is added to handle closing the window and resetting the user name. The visibility of the "Add Members to Channel" component is also toggled.

Note: This commit message follows the established convention of using the prefix "refactor" for code changes that improve or optimize existing functionality.