feat: rough popUp Overlay Structure for editing Msg´s implemented
The rough structure of a pop-up "editMsg" overlay was created within the chat history. This is used to load the text of an existing message into a text area using 2-way binding & reactive forms module. The author can then edit and save the message there. Additional functions will be implemented later.