feat: get / update Msg functions implemented
This commit includes:
- new get / update functions in the chat service to load & edit an existing message.
Using the getMessage function, the text of the message is returned and loaded into the ng form in the edit msg overlay.
Using the updateMessage function, the edited text which comes from the edit msg overlay and the ngform contained therein is updated in the document.

The following variables are required:
- mainCollectionName as a string
- the docref of the document which can be accessed
- the MessageId of the message to be edited