refactor/chore/fix/cleanup: private-note  & private-msg comp /  chat & firebase & communication service
A separate collection was created for the private notes. A privateNote.class was created. The PrivateChat.class was adjusted. The code in all affected components was adapted to the new situation.

Furthermore, 2 functions were added to Firebase & chatservice with which you can subscribe to and unsubscribe from all lists. This was done incorrectly within the service in ngOnit & onDestroy.