refactor: startNewPrivateChat optimized
The startNewPrivateChat function has been shortened and optimized. It first checks whether a chat exists between ChatCreator and ChatReceiver. If a chat already exists, the end user is redirected directly to the chat. If no chat exists, a new document is created in the Firestore database and then updated with the own docRef as ID. The end user is then redirected to the new private chat.