feat: build on searchService started & first functions added
This service receives search requests from all components and processes them. First, a searchUsersByName function was written as an observable. This processes the received search text and searches the Users collection and pushes the results into an array after comparing the search text. The contents of the array are rendered in the channel-member-selection, among other things. A pushSelectedUsersToArray function was also written. If you click on a user in the channel-member-selection, they are pushed into the selectedUser array. There is a query so that a member cannot be added twice. The last function written was removeSelectedUserFormArray. If you want to delete a user that has already been added from the selectedMembers array, then you can use this function.