refactor: Update searchService to handle member search in add-members-to-channel component
The searchService has been updated to handle member search functionality in the add-members-to-channel component. This includes refactoring the code related to adding members to the channel and moving the relevant code from the firebaseService to the searchService. Additionally, the searchUsersByName function has been modified to process the search text and search the Users collection, pushing the results into an array. The channel-member-selection component has been implemented to render the search results. The memberSearchActive flag has been added to toggle the display of search results. This refactor improves code organization and enhances the functionality of the application.