feat / refactor: setPalceholderName for channel/privateNote/privateChat implemented
The placeholder is first initialized by clicking on channel/privateChatList in the sidenav. If a user is already in a private chat/channel or private note and refreshes the browser there, the placeholder is reinitialized in the constructor of the components. A timeout is initially set for this so that the necessary lists of our subscriptions are fully loaded. There may be a better approach here.