feat: method implemented to update inital message wich gets messageAnswers
The message model has been adapted and expanded to include:
-answerCount, this is initially set to 0
-lastAnswer, is initially an empty string

When messageAnswers are created for a message, the initial message to which the answer refers is updated. The answerCount is increased by 1 and the time of the last messageAnswer written is always saved.

This enables the number of messageAnswers and the time of the last messageAnswer to be dynamically integrated into the chatHistory.