feat: findLatestMsgTime() implemented
findLatestMsgTime() is used as a helper function to update the lastAnswer field of the initial message when a messageAnswer is deleted. To do this, the msgAnswerList is searched, excluding the MsgAnswer to be deleted. It is then filtered by date and time to return the time of the last messageAnswer posted.