refactor: reactionService adding/deleting/updating messages with reactin object
The following has been implemented in the reaction service:
2 fixed reaction objects "ThumpUp +ThumpDown"

These fixed reactions are displayed as native emojis when hovering in the menu.

When a user clicks on it, the variable currentMessageId is saved with the msg.messageId. The "addSpecificReaction" method is then executed. It is given the messageId and one of the fixed objects.

The message doc is updated accordingly.

The updateReactionAmount method is then called with the "increase" operation, where the reactionID is used to check whether the specific reaction exists. If it exists, the amount is increased by 1. If the specific reaction does not exist, it is saved as an object in the message document in the reactions array.

The "handleReaction" method is called via the emojiPicker. It is given the messageId and the emojiEvent. The reaction object is created according to the event. The "addSpecificReaction" method is then called and takes over the logic described above.

The "deleteReaction" method is executed when you click on a reaction rendered under the chat message.