feat_ tagging lists for channels / privateChat & new-message
In channels, all channel members are displayed when you click on the tag button in the shared input field. When you click on the channel member, an @name is added to the message text.

In private chats, the chat receiver is displayed when you click on the tag button in the shared input field. When you click on the chat receiver, an @name is added to the message text.

In new messages, all users are displayed when you click on the tag button in the shared input field. When you click on the user, an @name is added to the message text.